Refractor rl question into one ask method without condition
